Preserve leading comments within jss-to-tss-react
